pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "qubership-git-system-follower"
version = "0.6.5"
description = "The package manager for Git providers"
authors = [
    { name = "Qubership" },
]
maintainers = [
    { name = "Vladislav Kishkin", email = "vladislav.kishkin01@gmail.com" },
]
license = "Apache-2.0"
license-files = ["LICENSE"]
readme = "README.md"
classifiers = [
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
]
keywords = [
    "package-manager", "state-management", "gear",
    "ci", "automation", "configuration", "infrastructure",
    "git", "gitlab",
    "devops", "sre",
    "python", "yaml"
]
requires-python = ">=3.10"
dependencies = [
    "click == 8.1.7",
    "colorama == 0.4.6",
    "cookiecutter == 2.6.0",
    "gitpython == 3.1.43",
    "python-gitlab == 4.9.0",
    "pyyaml == 6.0.2",
    "requests == 2.32.3",
    "packaging == 24.1",
    "oras == 0.2.24",
    "pluggy == 1.5.0",
    "outlify == 0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest == 8.3.4",
    "ruff == 0.9.1",
    "tox == 4.13.0",
    "vcrpy == 7.0.0",
]

docs = [
    "mkdocs-material == 9.6.11",
    "mike == 2.1.3",
]


[project.urls]
"Homepage" = "https://github.com/Netcracker/qubership-git-system-follower"
"Source" = "https://github.com/Netcracker/qubership-git-system-follower"
"Bug Tracker" = "https://github.com/Netcracker/qubership-git-system-follower/issues"
"Documentation" = "https://github.com/Netcracker/qubership-git-system-follower/blob/main/docs/docs_home.md"

[project.scripts]
git-system-follower = "git_system_follower.cli:cli"
gsf = "git_system_follower.cli:cli"


[tool.hatch.build.targets.wheel]
packages = ["git_system_follower"]

[tool.ruff]
line-length = 120
target-version = "py310"
output-format = "github"
exclude = [
  "tests/gears",
]

[tool.ruff.lint]
# See list of rules at: https://docs.astral.sh/ruff/rules/
select = ["E", "F", "W", "C90"]


[tool.pytest.ini_options]
testpaths = [
    "tests"
]

#[build-system]
#requires = ["poetry-core"]
#build-backend = "poetry.core.masonry.api"

##Optional [[tool.poetry.source]]
##Optional name = "gsf"
##Optional url = "REPLACE_WITH_CUSTOM_PYPI_REPO"
